renames "settings" files to "settings-panel"
