+++ src/components/index.tsx
@@ -5,7 +5,7 @@ import Board from './board';
 import RevisionHistory from './revision-history';
 
 import './index.css';
-import SettingsPanel from './settings';
+import SettingsPanel from './settings-panel';
 
 function App(): JSX.Element {
     const { toasts } = useToasterStore();
